CNN_LSTM/requirements.txt
scipy
numpy
soundfile
python-levenshtein
torch
torchelastic
visdom
wget
librosa
numba==0.43.0
llvmlite==0.32.1
tqdm
matplotlib
flask
sox
sklearn
soundfile
pytest
hydra-core==1.0.0rc1
google-cloud-storage
jupyter
ptflops